pyproject.toml
[project]
name = "investment-copilot"
version = "0.1.0"
description = "Investment copilot for the Polish stock market (GPW) with backtesting, powered by Groq."
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "MIT" }
authors = [{ name = "Investment Copilot" }]

dependencies = [
    "pydantic>=2.7",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "pandas>=2.2",
    "numpy>=1.26",
    "pyarrow>=16.0",
    "requests>=2.32",
    "feedparser>=6.0.11",
    "groq>=0.11",
    "typer>=0.12",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
]
gui = [
    "streamlit>=1.36",
    "plotly>=5.22",
]

[project.scripts]
invcopilot = "investment_copilot.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/investment_copilot"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
